refactor(tasks): dedupe checkbox helpers
