+++ ito-rs/crates/ito-domain/src/tasks/update.rs
@@ -5,57 +5,7 @@ use regex::Regex;
 
 use super::TaskStatus;
 
-fn is_checkbox_task_id_token(id: &str) -> bool {
-    let bytes = id.as_bytes();
-    if bytes.is_empty() {
-        return false;
-    }
-    if !bytes[0].is_ascii_digit() || !bytes[bytes.len() - 1].is_ascii_digit() {
-        return false;
-    }
-
-    let mut prev_dot = false;
-    for &b in bytes {
-        match b {
-            b'0'..=b'9' => {
-                prev_dot = false;
-            }
-            b'.' => {
-                if prev_dot {
-                    return false;
-                }
-                prev_dot = true;
-            }
-            _ => return false,
-        }
-    }
-    true
-}
-
-fn split_checkbox_task_label(s: &str) -> Option<(&str, &str)> {
-    let s = s.trim_start();
-    if s.is_empty() {
-        return None;
-    }
-
-    let bytes = s.as_bytes();
-    let mut split_at: Option<usize> = None;
-    for (i, &b) in bytes.iter().enumerate() {
-        if b == b' ' || b == b'\t' {
-            split_at = Some(i);
-            break;
-        }
-    }
-
-    let i = split_at?;
-    let (token, rest) = s.split_at(i);
-    let token = token.strip_suffix(':').unwrap_or(token);
-    let token = token.strip_suffix('.').unwrap_or(token);
-    if !is_checkbox_task_id_token(token) {
-        return None;
-    }
-    Some((token, rest.trim()))
-}
+use super::checkbox::split_checkbox_task_label;
 
 /// Update a checkbox-format task's status.
 ///
@@ -80,7 +30,10 @@ pub fn update_checkbox_task_status(
         }
     };
 
-    let mut lines: Vec<String> = contents.lines().map(|l| l.to_string()).collect();
+    let mut lines: Vec<String> = Vec::new();
+    for line in contents.lines() {
+        lines.push(line.to_string());
+    }
 
     // Prefer explicit ids when the task text starts with a numeric token (e.g. `1.1 First`).
     for line in &mut lines {
@@ -187,7 +140,10 @@ pub fn update_enhanced_task_status(
     let date = now.format("%Y-%m-%d").to_string();
     let updated_at_line = format!("- **Updated At**: {date}");
 
-    let mut lines: Vec<String> = contents.lines().map(|l| l.to_string()).collect();
+    let mut lines: Vec<String> = Vec::new();
+    for line in contents.lines() {
+        lines.push(line.to_string());
+    }
     let mut start_idx: Option<usize> = None;
     for (i, line) in lines.iter().enumerate() {
         if heading.is_match(line) {
@@ -238,7 +194,7 @@ pub fn update_enhanced_task_status(
                 lines.insert(end, updated_at_line);
                 lines.insert(end + 1, status_line);
             }
-            (Some(_), Some(_)) => {}
+            (Some(_status_idx), Some(_updated_idx)) => {}
         }
     }
